weird bug?

+++ code/lib/toml/src/main/scala/dev/librecybernetics/types/TOML.scala
@@ -4,4 +4,4 @@ enum TOML:
   case Comment(content: Predef.String)
   case String(content: Predef.String)
   case Integer(content: Int)
-  case Float(float: Float)
+  case Float(double: Double)
